requirements.txt
Faker==33.1.0
kafka-python==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
